+++ CHANGES.md
@@ -2,7 +2,10 @@
 
 ## 0.4.9 (2018-03-xx)
 
+* Wake payload reading task when data is available
+
 * Fix server keep-alive handling
+
 
 ## 0.4.8 (2018-03-12)